requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
wikipedia
langchain-text-splitters
langchain_openai
chromadb
sentence_transformers
langchain_huggingface
faiss_cpu
langchain_chroma
streamlit
langchain-groq
langchain-ollama
flask==2.3.3
python-dotenv==1.0.1
langchain-community==0.0.181
langchain-core==0.1.230
